requirements.txt
pyzbar==0.1.8
pillow==8.1.1